moved methods to classes
