+++ src/main/webclient/src/components/dashboard-page/components/chart-creator.tsx
@@ -1,6 +1,5 @@
 import {
   Button,
-  Center,
   Flex,
   Grid,
   GridItem,
@@ -20,11 +19,10 @@ import {
   chartType,
   getEmptyPreset,
 } from "../../../types/chartTemplate";
-import { Add, Info, InfoOutlined } from "@mui/icons-material";
+import { InfoOutlined } from "@mui/icons-material";
 import { NewCustomChartCreator } from "./new-custom-chart-creator";
 import { UIProps } from "../../../config/config";
-import { ChartData } from "@mantine/charts";
-import { saveAs } from "file-saver";
+import { localStorageKey, LocalStorageManager, FileSaver } from '../../../types/fileSaver';
 
 interface IChartCreator {
   model: DeviceModel;
@@ -32,28 +30,10 @@ interface IChartCreator {
 }
 
 export const ChartCreator = ({ model, devicesUptime }: IChartCreator) => {
-  const localStorageKey = "chartPresets";
-
-  const savePresetsToLocalStorage = (presets: ChartTemplate[]) => {
-    localStorage.setItem(
-      localStorageKey,
-      JSON.stringify(presets.map((p) => p.toJSON()))
-    );
-  };
-
-  const loadPresetsFromLocalStorage = (): ChartTemplate[] => {
-    const savedPresets = localStorage.getItem(localStorageKey);
-    // const removed = localStorage.removeItem(localStorageKey);
-    if (savedPresets) {
-      return JSON.parse(savedPresets).map((presetData: any) =>
-        ChartTemplate.fromJSON(presetData)
-      );
-    }
-    return [];
-  };
+  const localStorageKey: localStorageKey = "chartPresets";
 
   const [chartPresets, setChartPresets] = useState<ChartTemplate[]>(() => {
-    const presets = loadPresetsFromLocalStorage();
+    const presets = LocalStorageManager.loadPresetsFromLocalStorage(localStorageKey);
     console.log(presets);
     if (presets.length > 0) {
       return presets;
@@ -72,17 +52,6 @@ export const ChartCreator = ({ model, devicesUptime }: IChartCreator) => {
     ];
   });
 
-  const saveChartPresets = (templates: ChartTemplate[]) => {
-    const presetsJSON = JSON.stringify(
-      templates.map((preset) => preset.toJSON()),
-      null,
-      2
-    );
-    const blob = new Blob([presetsJSON], { type: "application/json" });
-    saveAs(blob, "chartPresets.json");
-    // savePresetsToLocalStorage(chartPresets);
-  };
-
   const addOrUpdatePreset = (newPreset: ChartTemplate) => {
     setChartPresets((prevPresets) => {
       const updatedPresets = [...prevPresets];
@@ -94,15 +63,15 @@ export const ChartCreator = ({ model, devicesUptime }: IChartCreator) => {
       } else {
         updatedPresets.push(newPreset);
       }
-      savePresetsToLocalStorage(updatedPresets);
-      saveChartPresets(updatedPresets);
+      LocalStorageManager.savePresetsToLocalStorage(localStorageKey, updatedPresets);
+      FileSaver.saveChartPresetsToJson(updatedPresets);
       console.log(updatedPresets);
       return updatedPresets;
     });
   };
 
   useEffect(() => {
-    const savedPresets = loadPresetsFromLocalStorage();
+    const savedPresets = LocalStorageManager.loadPresetsFromLocalStorage(localStorageKey);
     if (savedPresets.length > 0) {
       setChartPresets(savedPresets);
     }
